Linking to a commit in the `lug-template` repo instead of its latest ...
…state.

When I added more detail to this repository's repository, I added a link to the `lug-template` repo's latest
README. With that, if the `lug-template`'s README got updated, this link would automatically point to the
updated version. That seemed desirable for some reason, probably because sometimes the README gets fixed simply
to better reflect the current state of that template, and the template *seems* like it's in a good state. But I
now think that *today's* `lug-template` README is probably the one that will best describe these slides in
perpetuity. After all, these slides in this repo are very unlikely to change. So now I'm linking permanently to
the current state of the README as of today.